999+ move player script errors, I deleted the script from my branch because I can't edit
MissingComponentException: There is no 'Rigidbody2D' attached to the "Cat-1-Walk_0" game object, but a script is trying to access it.
You probably need to add a Rigidbody2D to the game object "Cat-1-Walk_0". Or your script needs to check if the component is attached before using it.
PlayerMovement.FixedUpdate () (at Assets/scripts/movePlayer.cs:101)